Rename file banner.vue to Banner.vue
