+++ pages/index.vue
@@ -38,7 +38,7 @@ import HeroBanner from '~/components/slices/hero-banner'
 import ImgTextColumnSlice from '~/components/slices/img-text-column'
 import SectionSlice from '~/components/slices/section'
 import PopInCampaigns from '~/components/pop-in-campaigns'
-import Banner from '~/components/slices/banner'
+import Banner from '@/components/slices/Banner'
 
 export default {
   components: {
